Moving Person class to AddressBook namespace - #254
